+++ views/confirmed.ejs
@@ -25,12 +25,11 @@
             </div>
         </nav>
         <section id="confirmation-section">
-            <img id="welcome-img" src="zeng-edited.jpg">
+            <img id="welcome-img" src="zeng2.jpg">
             <div id="welcome-message-h">
                 <h2>成功！</h2>
-                <h1>欢迎来到社区。</h1>
+                <h1>欢迎来到社区，<%= name %>!</h1>
             </div>
-
         </section>
         <footer>
             <%- include('./partials/footer.ejs') %>
