+++ views/home.ejs
@@ -16,10 +16,10 @@
                     target="_self"><span>家</span></a>
                 <a id="mobile-about-link" class="mobile-link" href="https://chineseflix.com/about"
                     target="_self"><span>关于我们</span></a>
-                <a id="mobile-login-link" class="mobile-link" href="https://chineseflix.com/user/login" target="_self"
-                    style="display: none;"><span>登录</span></a>
-                <a id="mobile-account-link" class="mobile-link" href="https://chineseflix.com/user/account"
-                    target="_self" style="display: none;"><span>帐户</span></a>
+                <a id="mobile-login-link" class="mobile-link" href="http://localhost:8080/login"
+                    target="_self"><span>登录</span></a>
+                <a id="mobile-account-link" class="mobile-link" href="https://chineseflix.com/account" target="_self"
+                    style="display: none;"><span>帐户</span></a>
                 <a id="mobile-contact-link" class="mobile-link" href="https://chineseflix.com/contact"
                     target="_self"><span>联系我们</span></a>
                 <a id="mobile-subtitle-help-link" class="mobile-link" href="https://chineseflix.com/subtitles"
