+++ views/login.ejs
@@ -25,33 +25,57 @@
             </div>
         </nav>
         <section id="login-signup-section">
+            <div id="login-img-parent">
+                <img src="zeng-edited.jpg">
+            </div>
             <div class="login-signup-form-container">
-                <h2>登录</h2>
-                <form id="login-form">
-
+                <h2 id="login-signup-h2">登录</h2>
+                <form id="login-form" action="/user/login" method="post">
                     <div id="login-input-container">
-                        <input type="email" id="login-email" placeholder="电子邮件" required>
-                        <input type="password" id="login-password" placeholder="密码" required>
+                        <input name="email" type="email" id="login-email" placeholder="电子邮件" required
+                            autocomplete="email">
+                        <input name="password" type="password" id="login-password" placeholder="密码" required
+                            autocomplete="current-password">
                     </div>
+                    <button id="login-button" type="submit">登录</button>
+                    <p>没有账户？ <span id="show-signup">注册</span></p>
 
-                    <button type="submit">登录</button>
-                    <p>沒有帳戶？ <a href="#" id="show-signup">报名</a></p>
+                    <!-- Display error message if there is one -->
+                    <% if (typeof error !=='undefined' ) { %>
+                        <div class="error-message" style="color: red;">
+                            <%= error %>
+                        </div>
+                        <% } %>
                 </form>
-                <form id="signup-form" style="display:none;">
-                    <h2>报名</h2>
-                    <input type="text" id="signup-username" placeholder="Username" required>
-                    <input type="email" id="signup-email" placeholder="Email" required>
-                    <input type="password" id="signup-password" placeholder="Password" required>
-                    <button type="submit">Sign Up</button>
-                    <p>Already have an account? <a href="#" id="show-login">Login</a></p>
+
+                <form id="signup-form" method="post" action="/user/signup">
+                    <div id="signup-input-container">
+                        <input id="signup-displayname" name="name" type="text" placeholder="输入您希望显示的全名"
+                            autocomplete="name" required>
+                        <input name="email" type="email" id="signup-email" placeholder="电子邮件" autocomplete="email"
+                            required>
+                        <input name="password" type="password" id="signup-password" minlength="8" maxlength="20"
+                            placeholder="创建一个安全的密码" required>
+                        <input name="confirmPassword" type="password" id="signup-confirm-password" placeholder="确认密码"
+                            autocomplete="new-password" required>
+                    </div>
+                    <button id="signup-button" type="submit">注册</button>
+                    <p>已经有账户？ <span id="show-login">登录</span></p>
+
+                    <!-- Display error message if there is one -->
+                    <% if (typeof error !=='undefined' ) { %>
+                        <div class="error-message" style="color: red;">
+                            <%= error %>
+                        </div>
+                        <% } %>
                 </form>
             </div>
         </section>
         <footer>
             <%- include('./partials/footer.ejs') %>
         </footer>
+        <%- include('./partials/events.ejs') %>
     </main>
-    <%- include('./partials/events.ejs') %>
 </body>
 
 </html>
