+++ views/partials/heading.ejs
@@ -10,7 +10,7 @@
     <a id="subtitle-help-link" href="https://chineseflix.com/subtitles" target="_self">
         字幕指南
     </a>
-    <a id="login-link" href="" target="_self" style="display: none;">登录</a>
+    <a id="login-link" href="http://localhost:8080/login" target="_self">登录</a>
     <a id="account-link" href="" target="_self" style="display: none;">帐户</a>
     <a id="contact-link" href="https://chineseflix.com/contact" target="_self">联系我们</a>
 </div>
